chore: #37 VisitedFeature 수정
